Split terminal input parity tests
