+++ src/adapters/terminal/input/tests.rs
@@ -20,6 +20,8 @@ use super::{EventSource, InputFailure, InputLane, InputMessage, translate};
 
 #[path = "tests/pointer.rs"]
 mod pointer;
+#[path = "tests/primary.rs"]
+mod primary;
 
 struct FakeSource {
     polls: VecDeque<io::Result<bool>>,
@@ -117,50 +119,6 @@ fn stalled_registry_event_source_becomes_a_typed_failure() {
     assert_eq!(failure, InputFailure::Unresponsive);
     lane.stop(ShutdownDeadline::after(Duration::from_secs(1)))
         .expect("input supervisor stops without the stalled reader");
-}
-
-#[test]
-fn command_and_meta_shortcuts_share_semantics() {
-    for modifier in [
-        KeyModifiers::CONTROL,
-        KeyModifiers::SUPER,
-        KeyModifiers::META,
-    ] {
-        let event = Event::Key(KeyEvent::new(KeyCode::Char('a'), modifier));
-        assert_eq!(translate(event), Some(UiInput::Key(UiKey::SelectAll)));
-    }
-}
-
-#[test]
-fn reserved_primary_chords_ignore_shifted_case_encoding() {
-    for modifier in [
-        KeyModifiers::CONTROL,
-        KeyModifiers::SUPER,
-        KeyModifiers::META,
-    ] {
-        for (lowercase, uppercase, expected) in [
-            ('a', 'A', UiKey::SelectAll),
-            ('c', 'C', UiKey::Copy),
-            ('x', 'X', UiKey::Cut),
-            ('v', 'V', UiKey::PasteClipboard),
-            ('d', 'D', UiKey::Duplicate),
-            ('q', 'Q', UiKey::Quit),
-            ('p', 'P', UiKey::PickerPrevious),
-            ('n', 'N', UiKey::PickerNext),
-            ('y', 'Y', UiKey::Redo),
-        ] {
-            for character in [lowercase, uppercase] {
-                assert_eq!(
-                    translate(Event::Key(KeyEvent::new(
-                        KeyCode::Char(character),
-                        modifier | KeyModifiers::SHIFT,
-                    ))),
-                    Some(UiInput::Key(expected)),
-                    "character {character:?}, modifier {modifier:?}"
-                );
-            }
-        }
-    }
 }
 
 #[test]
